feat(EventHandler): cache IDs of seen events in Redis to enforce idempotence
Most applications return the original result for subsequent calls to an
idempotent operation.  In our case, we just have to indicate that we've
already processed the event, since (a) the client already knows the
expected result of the operation and can apply it locally, and (b) in
the event of any conflict or drift it should catch up to the server
within the same (or at worst one additional) sync period.